Added new DController object.
DControlTransFunc defines a transfer function for the controller working
in an open loop. It recieves the input speed and position in the axes
space and compares it to the set point in the DController object.

DController object represents a motor torque controller for the
manipulator. It recieves speed and position as set point and input, then
it calculates the torque as set by a DControlTransFunc.

Both input and set point should be kept in sync with the output torque
signal.